deployment/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
pydantic-settings
dotenv
farm-haystack
awscli
awsebcli
